requirements.txt
# utilities
pydantic>=1.5.1,<1.6
loguru>=0.4.1,<0.5
PyYAML>=5.3.1,<5.4
Jinja2>=2.11,<2.12

# setuptools
pytest>=5.4,<5.5
